+++ src/ServiceStack.Redis/RedisSubscription.cs
@@ -10,7 +10,7 @@ namespace ServiceStack.Redis
     {
         private readonly IRedisNativeClient redisClient;
         private List<string> activeChannels;
-        public int SubscriptionCount { get; private set; }
+		public long SubscriptionCount { get; private set; }
         public bool IsPSubscription { get; private set; }
 
         private const int MsgIndex = 2;
